Move SimpleScope/SimpleType logic under Rector\Analyzer namespace
Group the PHPStan-free scope and type resolution under src/Analyzer:
- Rector\Analyzer\SimpleScope\{SimpleScope,SimpleScopeResolver}
- Rector\Analyzer\SimpleType\* (+ Contract\SimpleTypeInterface)
Tests mirror the layout under tests/Analyzer.
