+++ src/Analyzer/SimpleType/Contract/SimpleTypeInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Rector\SimpleType\Contract;
+namespace Rector\Analyzer\SimpleType\Contract;
 
 // PHPStan-free type contract, resolved by SimpleScope
 interface SimpleTypeInterface
